fix errors after review
