+++ src/utils/button-codes.js
@@ -1,1 +1,9 @@
-export const ESC_BUTTON_CODE = `Escape`;
+const ESC_BUTTON_CODE = `Escape`;
+
+export const isEscPressed = (evt) => {
+  if (evt.key === ESC_BUTTON_CODE) {
+    return true;
+  } else {
+    return false;
+  }
+};
