--- src/utils/filter-type.js
@@ -1,5 +1,0 @@
-export const FilterType = {
-  EVERYTHING: `EVERYTHING`,
-  FUTURE: `FUTURE`,
-  PAST: `PAST`
-};
